% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_helper.R
\name{are_all_close}
\alias{are_all_close}
\title{Check if two quantities are close enough to each other}
\usage{
are_all_close(v, w, abs_tol = 1e-06, rel_tol = 1e-06)
}
\description{
Check if two quantities are close enough to each other
}
