% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mle_pseudoinv.R
\name{glm_mle_pseudoinv}
\alias{glm_mle_pseudoinv}
\title{Implement the MLE finder via pseudo-inverse (using Cholesky decomposition)}
\usage{
glm_mle_pseudoinv(design, outcome)
}
\description{
Implement the MLE finder via pseudo-inverse (using Cholesky decomposition)
}
