% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_helper.R
\name{lm_log_likelihood}
\alias{lm_log_likelihood}
\title{Calculate the log-likelihood for a linear model}
\usage{
lm_log_likelihood(design, outcome, beta, noise_var = 1)
}
\description{
Calculate the log-likelihood for a linear model
}
