% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_helper.R
\name{lm_loglik_gradient}
\alias{lm_loglik_gradient}
\title{Calculate the gradient for a linear model}
\usage{
lm_loglik_gradient(design, outcome, beta, noise_var = 1)
}
\description{
Calculate the gradient for a linear model
}
