% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mle_pseudoinv.R
\name{lm_mle_pseudoinv}
\alias{lm_mle_pseudoinv}
\title{Implement the MLE finder via pseudo-inverse (using Cholesky decomposition)}
\usage{
lm_mle_pseudoinv(design, outcome)
}
\description{
Implement the MLE finder via pseudo-inverse (using Cholesky decomposition)
}
