% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_mle_BFGS.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the log-likelihood under a linear model}
\usage{
log_likelihood(design, outcome, beta, noise_var = 1)
}
\description{
Calculate the log-likelihood under a linear model
}
