% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_helper.R
\name{logit_hessian}
\alias{logit_hessian}
\title{Calculate the Hessian of the log-likelihood for a logistic regression model}
\usage{
logit_hessian(design, beta)
}
\description{
Calculate the Hessian of the log-likelihood for a logistic regression model
}
