% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_helper.R
\name{logit_log_likelihood}
\alias{logit_log_likelihood}
\title{Calculate the log-likelihood for a logistic regression model}
\usage{
logit_log_likelihood(design, outcome, beta)
}
\description{
Calculate the log-likelihood for a logistic regression model
}
