% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_helper.R
\name{logit_loglik_gradient}
\alias{logit_loglik_gradient}
\title{Calculate the gradient for a logistic regression model}
\usage{
logit_loglik_gradient(design, outcome, beta)
}
\description{
Calculate the gradient for a logistic regression model
}
