% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_mle_BFGS.R
\name{loglik_gradient}
\alias{loglik_gradient}
\title{Calculate the gradient under a linear model}
\usage{
loglik_gradient(design, outcome, beta, noise_var = 1)
}
\description{
Calculate the gradient under a linear model
}
