% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_finder.R
\name{mle_BFGS}
\alias{mle_BFGS}
\title{Implement the MLE finder via BFGS for linear and logistic models}
\usage{
mle_BFGS(design, outcome, model, ...)
}
\description{
Implement the MLE finder via BFGS for linear and logistic models
}
