% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_finder.R
\name{mle_newton}
\alias{mle_newton}
\title{Implement the MLE finder via Newton's method for logistic models}
\usage{
mle_newton(design, outcome, tol = NULL, max_iter = 50)
}
\description{
Implement the MLE finder via Newton's method for logistic models
}
