% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_finder.R
\name{mle_pseudoinv}
\alias{mle_pseudoinv}
\title{Implement the MLE finder via pseudo-inverse (using Cholesky decomposition) for linear models}
\usage{
mle_pseudoinv(design, outcome)
}
\description{
Implement the MLE finder via pseudo-inverse (using Cholesky decomposition) for linear models
}
